refactor(epoch-sync): move proof generation logic to epoch-manager
