+++ chain/epoch-manager/Cargo.toml
@@ -19,6 +19,7 @@ parking_lot.workspace = true
 primitive-types.workspace = true
 rand.workspace = true
 rand_hc.workspace = true
+rayon.workspace = true
 serde.workspace = true
 tracing.workspace = true
 # itertools has collect_vec which is useful in quick debugging prints
